Added Int field and cleaned up Err field
After casting Int to Int64 multiple times in log messages a basic Int
field was added to make this more transparent. The Int type in Go
matches the word length of the processor (32-bit or 64-bit) so it is
cast to an Int64 for simplicity in logging.

Also the Err field was changed so that it doesn't need a field name.
This was another annoyance as the field name was "error" in 99% of the
code logging errors, so it made sense to eliminate the field name
completely from the function and default it to "error"